+++ docs/examples/analysis/plot_efa_keller_massart.py
@@ -95,8 +95,9 @@ C = efa.transform()
 C.T.plot(title="EFA concentration")
 
 # %%
-# This ends the example ! The following line can be uncommented if no plot shows when running
-# the .py script
+# This ends the example ! The following line can be uncommented if no plot shows when
+# running the .py script
 
 # %%
+
 # scp.show()