+++ docs/examples/analysis/plot_pls.py
@@ -68,8 +68,9 @@ _ = pls.parityplot(
 _ = ax.legend(loc="lower right")
 
 # %%
-# This ends the example ! The following line can be uncommented if no plot shows when running
-# the .py script
+# This ends the example ! The following line can be uncommented if no plot shows when
+# running the .py script
 
 # %%
+
 # scp.show()